requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
googletrans==3.1.0a0
requests==2.31.0
google-cloud-translate==2.0.1